--- a/uberenv.py
+++ b/uberenv.py
@@ -796,13 +796,13 @@ class SpackEnv(UberEnv):
 
     def clean_build(self):
         if not self.opts["keep_stage"]:
-           # clean out any temporary spack build stages
-           cln_cmd = "spack/bin/spack clean "
-           res = sexe(cln_cmd, echo=True)
+            # clean out any temporary spack build stages
+            cln_cmd = "spack/bin/spack clean "
+            res = sexe(cln_cmd, echo=True)
 
-           # clean out any spack cached stuff
-           cln_cmd = "spack/bin/spack clean --all"
-           res = sexe(cln_cmd, echo=True)
+            # clean out any spack cached stuff
+            cln_cmd = "spack/bin/spack clean --all"
+            res = sexe(cln_cmd, echo=True)
 
         # check if we need to force uninstall of selected packages
         if self.opts["spack_clean"]: