--- a/uberenv.py
+++ b/uberenv.py
@@ -230,14 +230,6 @@ def parse_args():
                       default=False,
                       help="Only download and setup Spack. No further Spack command will be run.")
 
-    # option to keep staged directories from previous Uberenv runs
-    parser.add_option("--keep-stage",
-                      action="store_true",
-                      dest="keep_stage",
-                      default=False,
-                      help="Keep stage directories from previous Uberenv runs;"
-                           "does not override --clean.")
-    
     ###############
     # parse args
     ###############
@@ -795,14 +787,10 @@ class SpackEnv(UberEnv):
 
 
     def clean_build(self):
-        if not self.opts["keep_stage"]:
-            # clean out any temporary spack build stages
-            cln_cmd = "spack/bin/spack clean "
-            res = sexe(cln_cmd, echo=True)
-
-            # clean out any spack cached stuff
-            cln_cmd = "spack/bin/spack clean --all"
-            res = sexe(cln_cmd, echo=True)
+        # clean out any spack cached stuff (except build stages, downloads, &
+        # spack's bootstrapping software)
+        cln_cmd = "spack/bin/spack clean --misc-cache --failures --python-cache"
+        res = sexe(cln_cmd, echo=True)
 
         # check if we need to force uninstall of selected packages
         if self.opts["spack_clean"]: