EDIT: Instead of making the changes described below, I followed
@white238's suggestion and changed the arguments to one of
the `spack clean` calls in Uberenv.

---------------

This commit adds a "--keep-stage" flag that disables the initial
`spack clean` and `spack clean --all` invocations when running
Uberenv. The name of this flag was chosen to evoke the same semantics
as the `--keep-stage` flag in `spack install`, `spack create`,
etc. However, this flag will *not* override the `--clean` flag of the
Uberenv CLI; I assume that if a user calls Uberenv with that flag,
they do intend to forcibly remove installations of previous packages
and their dependents.

The use case for this flag is the following: In one project I work on,
Uberenv is invoked repeatedly with different specs to install several
versions of third-party libraries. On this project, I would like to
keep the staged directories after the first invocation so that my
colleagues can drop into the source files of third-party libraries in
a debugger.