@white238 2021-06-23
It was my understanding that we were utilizing the `spack clean` command to remove the temporary and cache files but the comment clearly states otherwise.  I would prefer to leave the staging directories by default.

After reading the documentation here:

https://spack.readthedocs.io/en/latest/packaging_guide.html?highlight=clean#spack-clean

I would prefer to just lock down the command to something more like `spack clean --misc-cache --failures --python-cache` and then not have another option.

Am I missing a side effect this may cause, @cyrush ?