@goxberry 2021-07-15
@kennyweiss @white238 @cyrush Is there anything I can do to help this PR land in Uberenv? One of the projects I work on is waiting for this PR to land. I can reconfigure our uberenv git submodule instance to point to my fork, but we have several junior developers working on that project, so if I can avoid that reconfiguration, it would save me the trouble of having to debug their setups with them.