@cyrush 2021-07-15
My perspective: aggressive cleaning (`spack clean --all`) help avoid surprises when you rerun uberenv. 

I am fine with an option to keep things around -- but I think it should be an option, I don't think we should remove the clean --all.