@cyrush 2021-07-15
Ok, I am a little confused: Don't we just need `--keep-stage` on the ``install`` command?

That seems totally orthogonal from the clean settings?