@white238 2021-07-15
To my understanding, `spack clean --all` removes previous stage directories.

https://spack.readthedocs.io/en/latest/command_index.html#spack-clean

This is the equivalent to something like `spack clean --all --keep-stage` but that doesnt exist.