@white238 2021-07-15
> 
> 
> Ok, so if you run uberenv -- everything installs fine, then you run uberenv again it will clean up the stages?
> 
> I think you still need --keep-stage for successful builds? We aren't calling clean after install.

correct