@cyrush 2021-07-15
Ok, lets aim for the clean change -- but I still think we need `--keep-stage`?

I think Spack's behavior has changed over time (added new things that are cached, etc)

It would help to have a few test runs shared so we can verify everything works like we expect.