@goxberry 2021-07-16
> I think you still need `--keep-stage` for successful builds? We aren't calling clean after install.

We're using `dev-build` mode, and `spack dev-build` does not have a `--keep-stage` option. You're right that `--keep-stage` would need to be added to `spack install` invocations.