@goxberry 2021-07-16
> It would help to have a few test runs shared so we can verify everything works like we expect.

Is there a test suite somewhere?