@cyrush 2021-07-16
Let's make the change.


Spack's clean behavior has changed over time.
`spack clean` did not always equate to `spack clean --all`, however that is now the case. 

This was throwing me off.  Clean `--all` wasn't even removing the stages in the older version I was testing. 
I also think that dev-build might have different rules. 

To move forward, I was just looking for an example to make sure we have a handle on the current spack behavior. 
We *do* need `--keep-stage` for any non dev build cases. I added a ticket for this: #68.

Here is an example of the difference with when installing.
(with `--keep-stage`)

```
[harrison37@elder spack (develop)]$ bin/spack clean --all 
==> Removing all temporary build stages
==> Removing cached downloads
==> Removing install failure marks
==> Removing cached information on repositories
==> Removing python cache files
==> Removing software in "/Users/harrison37/.spack/bootstrap/store"
[harrison37@elder spack (develop)]$ bin/spack install --keep-stage libxml2 
[+] /Users/harrison37/Work/spack/opt/spack/darwin-mojave-skylake/apple-clang-11.0.0/libiconv-1.16-2py5rvx5z25udmw3g5463rimshfkqqxa
[+] /Users/harrison37/Work/spack/opt/spack/darwin-mojave-skylake/apple-clang-11.0.0/pkgconf-1.7.4-riy5ze5ht5ch2ewfjpkgocdmr7iplxc5
[+] /Users/harrison37/Work/spack/opt/spack/darwin-mojave-skylake/apple-clang-11.0.0/xz-5.2.5-r6r6ntbbrxa4vwfifcrbpv3cbc6tfc7y
==> Installing zlib-1.2.11-vehd5dsoagzcexij7zkjr7tb7que36yw
==> No binary for zlib-1.2.11-vehd5dsoagzcexij7zkjr7tb7que36yw found: installing from source
==> Warning: microarchitecture specific optimizations are not supported yet on mixed compiler toolchains [check apple-clang@11.0.0 for further details]
==> Fetching https://mirror.spack.io/_source-cache/archive/c3/c3e5e9fdd5004dcb542feda5ee4f0ff0744628baf8ed2dd5d66f8ca1197cb1a1.tar.gz
==> No patches needed for zlib
==> zlib: Executing phase: 'install'
==> zlib: Successfully installed zlib-1.2.11-vehd5dsoagzcexij7zkjr7tb7que36yw
  Fetch: 3.21s.  Build: 4.19s.  Total: 7.39s.
[+] /Users/harrison37/Work/spack/opt/spack/darwin-mojave-skylake/apple-clang-11.0.0/zlib-1.2.11-vehd5dsoagzcexij7zkjr7tb7que36yw
==> Installing libxml2-2.9.10-mrq6i4jb2t5zoozreokxqoeixhrc5l4g
==> No binary for libxml2-2.9.10-mrq6i4jb2t5zoozreokxqoeixhrc5l4g found: installing from source
==> Warning: microarchitecture specific optimizations are not supported yet on mixed compiler toolchains [check apple-clang@11.0.0 for further details]
==> Fetching https://mirror.spack.io/_source-cache/archive/aa/aafee193ffb8fe0c82d4afef6ef91972cbaf5feea100edc2f262750611b4be1f.tar.gz
==> Fetching https://mirror.spack.io/_source-cache/archive/96/96151685cec997e1f9f3387e3626d61e6284d4d6e66e0e440c209286c03e9cc7.tar.gz
==> Moving resource stage
	source : /var/folders/wj/jjhjt_0d3kg1vmk56d5bq_38001kr8/T/harrison37/spack-stage/resource-xmlts-mrq6i4jb2t5zoozreokxqoeixhrc5l4g/spack-src/
	destination : /var/folders/wj/jjhjt_0d3kg1vmk56d5bq_38001kr8/T/harrison37/spack-stage/spack-stage-libxml2-2.9.10-mrq6i4jb2t5zoozreokxqoeixhrc5l4g/spack-src/xmlconf
==> Ran patch() for libxml2
==> libxml2: Executing phase: 'autoreconf'
==> libxml2: Executing phase: 'configure'
==> libxml2: Executing phase: 'build'
==> libxml2: Executing phase: 'install'
==> libxml2: Successfully installed libxml2-2.9.10-mrq6i4jb2t5zoozreokxqoeixhrc5l4g
  Fetch: 8.25s.  Build: 57.06s.  Total: 1m 5.32s.
[+] /Users/harrison37/Work/spack/opt/spack/darwin-mojave-skylake/apple-clang-11.0.0/libxml2-2.9.10-mrq6i4jb2t5zoozreokxqoeixhrc5l4g
[harrison37@elder spack (develop)]$ bin/spack location zlib
/var/folders/wj/jjhjt_0d3kg1vmk56d5bq_38001kr8/T/harrison37/spack-stage/spack-stage-zlib-1.2.11-vehd5dsoagzcexij7zkjr7tb7que36yw/spack-src
[harrison37@elder spack (develop)]$ bin/spack location libxml2
/var/folders/wj/jjhjt_0d3kg1vmk56d5bq_38001kr8/T/harrison37/spack-stage/spack-stage-libxml2-2.9.10-mrq6i4jb2t5zoozreokxqoeixhrc5l4g/spack-src
[harrison37@elder spack (develop)]$ bin/spack clean
==> Removing all temporary build stages
[harrison37@elder spack (develop)]$ bin/spack location zlib
==> Error: Source directory does not exist yet. Run this to create it:
  spack stage zlib
[harrison37@elder spack (develop)]$ bin/spack location libxml2
==> Error: Source directory does not exist yet. Run this to create it:
  spack stage libxml2
```


(without `--keep-stage`)
```
[harrison37@elder spack (develop)]$ bin/spack uninstall --dependents zlib 
==> The following packages will be uninstalled:

    -- darwin-mojave-skylake / apple-clang@11.0.0 -------------------
    mrq6i4j libxml2@2.9.10  vehd5ds zlib@1.2.11

==> Do you want to proceed? [y/N] y
==> Successfully uninstalled libxml2@2.9.10%apple-clang@11.0.0~python arch=darwin-mojave-skylake/mrq6i4j
==> Successfully uninstalled zlib@1.2.11%apple-clang@11.0.0+optimize+pic+shared arch=darwin-mojave-skylake/vehd5ds
[harrison37@elder spack (develop)]$ bin/spack install libxml2 
[+] /Users/harrison37/Work/spack/opt/spack/darwin-mojave-skylake/apple-clang-11.0.0/libiconv-1.16-2py5rvx5z25udmw3g5463rimshfkqqxa
[+] /Users/harrison37/Work/spack/opt/spack/darwin-mojave-skylake/apple-clang-11.0.0/pkgconf-1.7.4-riy5ze5ht5ch2ewfjpkgocdmr7iplxc5
[+] /Users/harrison37/Work/spack/opt/spack/darwin-mojave-skylake/apple-clang-11.0.0/xz-5.2.5-r6r6ntbbrxa4vwfifcrbpv3cbc6tfc7y
==> Installing zlib-1.2.11-vehd5dsoagzcexij7zkjr7tb7que36yw
==> No binary for zlib-1.2.11-vehd5dsoagzcexij7zkjr7tb7que36yw found: installing from source
==> Warning: microarchitecture specific optimizations are not supported yet on mixed compiler toolchains [check apple-clang@11.0.0 for further details]
==> Using cached archive: /Users/harrison37/Work/spack/var/spack/cache/_source-cache/archive/c3/c3e5e9fdd5004dcb542feda5ee4f0ff0744628baf8ed2dd5d66f8ca1197cb1a1.tar.gz
==> No patches needed for zlib
==> zlib: Executing phase: 'install'
==> zlib: Successfully installed zlib-1.2.11-vehd5dsoagzcexij7zkjr7tb7que36yw
  Fetch: 0.00s.  Build: 4.64s.  Total: 4.65s.
[+] /Users/harrison37/Work/spack/opt/spack/darwin-mojave-skylake/apple-clang-11.0.0/zlib-1.2.11-vehd5dsoagzcexij7zkjr7tb7que36yw
==> Installing libxml2-2.9.10-mrq6i4jb2t5zoozreokxqoeixhrc5l4g
==> No binary for libxml2-2.9.10-mrq6i4jb2t5zoozreokxqoeixhrc5l4g found: installing from source
==> Warning: microarchitecture specific optimizations are not supported yet on mixed compiler toolchains [check apple-clang@11.0.0 for further details]
==> Using cached archive: /Users/harrison37/Work/spack/var/spack/cache/_source-cache/archive/aa/aafee193ffb8fe0c82d4afef6ef91972cbaf5feea100edc2f262750611b4be1f.tar.gz
==> Using cached archive: /Users/harrison37/Work/spack/var/spack/cache/_source-cache/archive/96/96151685cec997e1f9f3387e3626d61e6284d4d6e66e0e440c209286c03e9cc7.tar.gz
==> Moving resource stage
	source : /var/folders/wj/jjhjt_0d3kg1vmk56d5bq_38001kr8/T/harrison37/spack-stage/resource-xmlts-mrq6i4jb2t5zoozreokxqoeixhrc5l4g/spack-src/
	destination : /var/folders/wj/jjhjt_0d3kg1vmk56d5bq_38001kr8/T/harrison37/spack-stage/spack-stage-libxml2-2.9.10-mrq6i4jb2t5zoozreokxqoeixhrc5l4g/spack-src/xmlconf
==> Ran patch() for libxml2
==> libxml2: Executing phase: 'autoreconf'
==> libxml2: Executing phase: 'configure'
==> libxml2: Executing phase: 'build'
==> libxml2: Executing phase: 'install'
==> libxml2: Successfully installed libxml2-2.9.10-mrq6i4jb2t5zoozreokxqoeixhrc5l4g
  Fetch: 0.03s.  Build: 56.67s.  Total: 56.70s.
[+] /Users/harrison37/Work/spack/opt/spack/darwin-mojave-skylake/apple-clang-11.0.0/libxml2-2.9.10-mrq6i4jb2t5zoozreokxqoeixhrc5l4g
[harrison37@elder spack (develop)]$ bin/spack location zlib 
==> Error: Source directory does not exist yet. Run this to create it:
  spack stage zlib
```